Add prompt warning on destructive query in `watch` command
A refactor was required to being able to import
`confirm_destructive_query` in `packages/special/iocommands.py`.

The tests moved have lost the `dbtest` decorator as it wasn't needed,
they works fine without a database connection.
